fix tests

+++ test/QField.jl
@@ -257,8 +257,8 @@ end
     a = QAdd([ϕ, ψ])
     b = QAdd([QMul(1, [ϕ]), QMul(1, [ψ])])
     @test isequal(a, b)
-    @test typeof(a + b) == QAdd{Float64}
-    @test typeof(a * b) == QAdd{Float64}
+    @test typeof(a + b) == QAdd{Int64}
+    @test typeof(a * b) == QAdd{Int64}
     @test typeof(convert(QAdd{Float64}, a)) == QAdd{Float64}
     @test typeof(convert(QMul{Float64}, QMul(1, [ϕ]))) == QMul{Float64}
 end
